+++ src/Codeception/Module/Laravel.php
@@ -7,18 +7,16 @@ namespace Codeception\Module;
 use Closure;
 use Codeception\Configuration;
 use Codeception\Exception\ModuleConfigException;
-use Codeception\Exception\ModuleException;
 use Codeception\Lib\Connector\Laravel as LaravelConnector;
 use Codeception\Lib\Framework;
 use Codeception\Lib\Interfaces\ActiveRecord;
 use Codeception\Lib\Interfaces\PartedModule;
 use Codeception\Lib\ModuleContainer;
+use Codeception\Module\Laravel\InteractsWithAuthentication;
 use Codeception\Subscriber\ErrorHandler;
 use Codeception\TestInterface;
 use Codeception\Util\ReflectionHelper;
 use Exception;
-use Illuminate\Contracts\Auth\Authenticatable;
-use Illuminate\Contracts\Auth\Factory as AuthContract;
 use Illuminate\Contracts\Console\Kernel;
 use Illuminate\Contracts\Routing\UrlGenerator;
 use Illuminate\Contracts\Session\Session;
@@ -132,6 +130,8 @@ use function is_array;
  */
 class Laravel extends Framework implements ActiveRecord, PartedModule
 {
+    use InteractsWithAuthentication;
+
     /**
      * @var Application
      */
@@ -789,81 +789,6 @@ class Laravel extends Framework implements ActiveRecord, PartedModule
         if (! is_null($errorMessage)) {
             $this->assertStringContainsString($errorMessage, $viewErrorBag->first($field));
         }
-    }
-
-    /**
-     * Set the currently logged in user for the application.
-     * Takes either an object that implements the User interface or
-     * an array of credentials.
-     *
-     * ``` php
-     * <?php
-     * // provide array of credentials
-     * $I->amLoggedAs(['username' => 'jane@example.com', 'password' => 'password']);
-     *
-     * // provide User object
-     * $I->amLoggedAs( new User );
-     *
-     * // can be verified with $I->seeAuthentication();
-     * ```
-     * @param Authenticatable|array $user
-     * @param string|null $guardName The guard name
-     */
-    public function amLoggedAs($user, ?string $guardName = null): void
-    {
-        /** @var AuthContract $auth */
-        $auth = $this->app['auth'];
-
-        $guard = $auth->guard($guardName);
-
-        if ($user instanceof Authenticatable) {
-            $guard->login($user);
-            return;
-        }
-
-        $this->assertTrue($guard->attempt($user), 'Failed to login with credentials ' . json_encode($user));
-    }
-
-    /**
-     * Logout user.
-     */
-    public function logout(): void
-    {
-        $this->app['auth']->logout();
-    }
-
-    /**
-     * Checks that a user is authenticated.
-     * You can specify the guard that should be use as second parameter.
-     *
-     * @param string|null $guard
-     */
-    public function seeAuthentication($guard = null): void
-    {
-        /** @var AuthContract $auth */
-        $auth = $this->app['auth'];
-
-        $auth = $auth->guard($guard);
-
-        $this->assertTrue($auth->check(), 'There is no authenticated user');
-    }
-
-    /**
-     * Check that user is not authenticated.
-     * You can specify the guard that should be use as second parameter.
-     *
-     * @param string|null $guard
-     */
-    public function dontSeeAuthentication(?string $guard = null): void
-    {
-        /** @var AuthContract $auth */
-        $auth = $this->app['auth'];
-
-        if (is_string($guard)) {
-            $auth = $auth->guard($guard);
-        }
-
-        $this->assertNotTrue($auth->check(), 'There is an user authenticated');
     }
 
     /**
